enterprises/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg  # for PostgreSQL
aiomysql  # for MySQL
aiosqlite  # for SQLite
pydantic
celery[zstd,msgpack,redis]==5.4.0
flower
pyyaml
passlib
python-jose[cryptography]
python-multipart
email-validator
docker
aiohttp